+++ src/atoms/index.tsx
@@ -2,3 +2,4 @@ export { loggedAtom } from './AtomContainer';
 export { editProfileModalAtom } from './AtomContainer';
 export { userWithdrawalModalAtom } from './AtomContainer';
 export { removeCommentModalAtom } from './AtomContainer';
+export { commentIdxAtom } from './AtomContainer';
